Devex 1187: Add new joi validation entity so we can transition entities one at a time to new validation error message pattern, implement with case deadline and add test coverage
